Don't assume that RingBuffer::Create always returns something.
There are a couple of places in tv_play.cpp where the result of this
call is immediately assigned into a structure.  Many of the routines
accessing that field of the structure check for non-NULL before using
it.  This change adds checks to protect the rest of code that accesses
that field.