Use FFmpeg to convert external subtitles to utf-8.
This requires examining the file contents before initializing the
FFmpeg subtitle codec.  If the file encoding is already utf-8, no
conversion is required.  If not, then the codec is initialized to
convert from user specified encoding to utf-8.  The encoding
conversion is handled integral to FFmpeg using the system iconv
library.