Revert "Add custom targets to rebuild mythtv or mythplugins"
On case insensitive file systems, the 'MythTV' target holding the
instructions for building the mythtv directory conflicts with the
'mythtv' target for easy recompiling of said directory.  Remove the
recently added 'mythtv'/'mythplugins' targets until better names can
be found.  The names 'myth' and 'plugins' were considered, but the
former conflicts with the required name 'myth' for the libs/libmyth
directory (the 'lib' prefix is added automatically when the library is
built, turning the name 'myth' into the required library name
'libmyth').

Until workable names are found, recompiling a sub-target will need to
be done with:

  cmake --build <builddir>/MythTV-prefix/src/MythTV-build/

or:

  cd <builddir>/MythTV-prefix/src/MythTV-build/
  cmake --build .

This reverts commit 5e5ef57071c6c57d5279de744b6c5e7b4e09d09f.