tidy: Quiet warning about bit shifting in bitreader.h.
The clang-tidy core.BitwiseShift pointed out undefined behavior
refilling from cache.  I cannot find an obvious problem here, but the
calculation depends on state from outside the function.  Unless code
elsewhere is wrong, the shift variable should be restricted to the
range [0..63] and the bits variable should be restricted to the range
[0..7].  Tidy is complaining that the clause "1<<bits" is overflowing
an integer, which isn't possible given these restrictions.

https://clang.llvm.org/docs/analyzer/checkers.html#core-bitwiseshift