tidy: Fix standard library misuse warning.
The clang-tidy analyzer is complaining that the getsockname and
getpeername library functions are being called with a first argument
of -1.  This check is actually happening, its just opaque to the
compiler.  Move the trivial isValid() function into the object
declaration so that the compiler fully understands the code execution
paths.

https://clang.llvm.org/docs/analyzer/checkers.html#unix-stdclibraryfunctions