cmake: Android builds now work with recent Qt versions.
The fix for Qt bug https://bugreports.qt.io/browse/QTBUG-98974 changed
the return values of a couple of the qfileinfo functions.  This bug
was committed to mainline Qt in 6.5.0, and back-ported to 6.4.3, 6.2.8,
and 5.15.13.  What now happens is that calling QFileInfo::fileName()
now returns "assets:/filename" instead of just "filename".

There are two different fixes here.

1) For the font manager, the code is walking a set of files in the
current directory.  Instead of having it return a list of filenames
and building absolute paths by hand, just have it return a list of
QFileInfo structures and use the absolute pathname in each of those.
This circumvents the problems caused by the fix for QTBUG-98974.

2) For the theme manager, the code is getting a list of filenames in a
directory, and then searching for those same filenames in other
directories.  Its not possible to circumvent the problem in this case,
so we need to explicitly strip the file system name ("assets:/") from
the returned filename.