Tweak initial creation of the "mythconverg" database. (1)
This change adds quotes to all table and column names.  This should
prevent problems when databases add keywords that conflict with
existing column names used by MythTV.  It also matches the default
output from mysqldump, making it easier to update the function that
creates the initial database.

These changes are best viewed with the "--word-diff" argument to the
"git diff" command.