Tweak initial creation of the "mythconverg" database. (2)
This change removes the quotes around numbers when specifying default
values in a CREATE TABLE statement.  The field type is declared as a
number earlier in the same line, so all quoting a number does is force
mysql to read it as a string and then convert it to a number, instead
of just reading it as a number.  This also matches the default output
from mysqldump, making it easier to update the function that creates
the initial database.

These changes are best viewed with the "--word-diff" argument to the
"git diff" command.

Note: This is from mysqldump against a mariadb database.  Mysqldump
against a mysql database still puts quotes around the numbers, but it
has other issues making the mariadb dump preferable.