Tweak initial creation of the "mythconverg" database. (3)
This change replaces uses of "CURRENT_TIMESTAMP" with
"current_timestamp()".  According to the mysql documentation, the two
are identical, but the latter matches the default output from
mysqldump, making it easier to update the function that creates the
initial database.

Note: This is from mysqldump against a mariadb database.  Mysqldump
against a mysql database still uses the macro form for this statement,
but it has other issues making the mariadb dump preferable.