Upgrade initial creation of the "mythconverg" database to version 1382.
Upgrade the initial database creation code to generate version 1382 of
the database (instead of version 1307).  This was done by deleting the
mythconverg database from mysql, creating a new instance of the
initial database using the command "mythbackend --upgradedbonly", and
then dumping its contents with mysqldump.  The InitializeMythSchema
function was then patched to match the output of mysqldump.

The changes in the mysqldump invocation (the sed commands) are to
remove mariadb version specific utf8 references (utf8mb3_xxx) and
replace them with generic utf8 references.  Mysql/mariadb will use its
default multibyte support (three or four bytes) when it encounters
these generic utf8 declarations.

Note: The data is from mysqldump against a mariadb database for two
reasons.  1) Mariadb maintains the display widths specified during
table creation, whereas mysql strips them from all integer based
number declarations.  The only exception is that mysql keeps the
display width for a declaration of tinyint(1).  2) Mariadb is what is
present on all my systems.  I did install mysql on a VM and test
against it to ensure that mythbackend can still initialize the db on a
mysql server, and that it produces the same results on both mariadb
and mysql.