cmake: Prevent weird errors from cmake configuration.
If the configuration of a project or subproject exits before the cmake
project() function is called, you get a very weird error about the
version of ninja being too low (< 1.3) regardless of the installed
version of ninja.  The android subproject is elsewhere protected from
being created if the ANDROID variable isn't set, so remove that test
here.  Just in case.  So if you do somehow invoke it when you
shouldn't you don't waste time chasing false clues.  :-)