tidy: Eliminate magic return value from selectDrivePopup.
Clang-tidy pointed out this file as casting integers to pointers,
which conceals information from the optimizer if the integer was
originally cast from a pointer.

In this case the pointer value of -1 is used as a magic value so that
the function can have a ternary result: drive, no drive, or canceled.
Add another argument to the function to pass out whether the user
selection was canceled, and make the function return value a simple
pointer instead of a ternary result.