cmake/android: Consolidate Qt6 android-34 platform declaration.
Qt6 requires a minimum android sdk API level of 26 (Android 8.0).
Compiling Qt6 against this level throws a variety of errors, most
complaining that API 33 is needed, and one complaining that API 34 is
needed.  Given that Google has also stated a requirement for API 34
for all new apps (as of 2024-08-31), set android-34 as the target for
all Qt6 builds.