Indent code for future changes from C to C++ error handling.
This commit contains no logic changes, only indentation changes.
There are a couple of places where a set of braces was also
added. This quiets compiler warnings that the newly indented code
wasn't protected by an "if" statement even though it appeared that way
because of the indentation.

This commit should be viewed with "git diff -w".