Fail HLS stream parsing the first time TARGETDURATION fails.
In certain cases the code performs an early targeted check for the
presence of a m3u8 TARGETDURATION entry.  If it finds such an entry
but parsing of the entry value fails, the current code continues
processing.  Shortly thereafter while processing every line in the
m3u8 file, it will process the TARGETDURATION entry again and this
time fail out because it can't parse the entry value.  Change the code
to bail out after the first failed parse instead of waiting for the
second failure.