Convert "do" loops to "while" loops. (libs/libs/libmythupnp/msocketdevice_unix.cpp)
Using a "while" loop is easier to understand because the exit
condition of the loop is stated up front.

Note: There doesn't appear to be any code path to reach this
function. The only calls to MSocketDevice::accept are from other
accept functions in the libmythupnp directory.  The only code that
creates an MSocketDevice (or MMulticastSocketDevice or
MBroadcastSocketDevice which derive from that class) is also in the
libmythupnp directory.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/avoid-do-while.html