Add some debugging code to empty catch() statements.
This should allow easier diagnosis of of thy an error was thrown.
Errors in the standard libraries will always throw at item derived
from a std::exception.

The MHEngine::FindObject function throws a char* if it can't find an
object (an expected case) so that the caller move on to the next
object.