mheg: Null terminate char version of MHOctetString.
The code has always allocated space for a null terminator but never
written it into the data structure.  This isn't a problem unless the
code uses the Bytes() function to get a pointer to the data and then
processes the data as a null terminated string.  Most of the existing
code uses the Size() function instead of the null terminator, so this
hasn't been a problem.