tidy: Fix libmythmpeg2 warnings about implicit widening of multiplication results.
These functions perform a lot of memory address calculations.  Cast
the widths used (8 and 4) to a ptrdiff_t type so that the calculations
are performed in a register of the proper width for a pointer, instead
of being calculated in a smaller width register and widened after the
fact.  Given the small sizes of these multiplications, this should all
be semantic sugar to silence the warnings, but it allows the warning
to now be enabled for the entire code base.