tidy: Fix vizualization warnings about implicit widening of multiplication results.
Change a couple of loop variables from int to ptrdiff_t so that the
multiplications are performed in a register of the proper width for a
pointer, instead of being calculated in a smaller width register and
widened after the fact.  This allows this warning to now be enabled
for the entire code base.